+++ src/Handler/HeaderProcessor.php
@@ -44,7 +44,7 @@ final class HeaderProcessor
             throw new \RuntimeException('HTTP status code missing from header data');
         }
 
-        if (!\preg_match('/^\d{3}$/', $status)) {
+        if (!\preg_match('/^[1-5]\d{2}$/', $status)) {
             throw new \RuntimeException('HTTP status code is invalid');
         }