+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Please refer to [UPGRADING](UPGRADING.md) guide for upgrading to a major version
 
 ## 7.11.0 - Upcoming
 
+### Changed
+
+- Adjusted `guzzlehttp/psr7` version constraint to `^2.10`
+
 ### Fixed
 
 - Improve invalid response handling across handlers