+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 Please refer to [UPGRADING](UPGRADING.md) guide for upgrading to a major version.
 
-## 7.10.1 - Upcoming
+## 7.10.1 - 2026-05-19
 
 ### Fixed